Move abstract search request to requests folder
