+++ src/Http/Requests/Announcements/SearchAnnouncementsRequest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Perscom\Http\Requests\Announcements;
 
-use Perscom\RequestType\AbstractSearchRequest;
+use Perscom\Http\Requests\AbstractSearchRequest;
 
 class SearchAnnouncementsRequest extends AbstractSearchRequest
 {
