+++ src/Http/Requests/Qualifications/SearchQualificationsRequest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Perscom\Http\Requests\Qualifications;
 
-use Perscom\RequestType\AbstractSearchRequest;
+use Perscom\Http\Requests\AbstractSearchRequest;
 
 class SearchQualificationsRequest extends AbstractSearchRequest
 {
