+++ src/Http/Requests/Ranks/SearchRanksRequest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Perscom\Http\Requests\Ranks;
 
-use Perscom\RequestType\AbstractSearchRequest;
+use Perscom\Http\Requests\AbstractSearchRequest;
 
 class SearchRanksRequest extends AbstractSearchRequest
 {
