+++ src/Http/Requests/Specialties/SearchSpecialtiesRequest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Perscom\Http\Requests\Specialties;
 
-use Perscom\RequestType\AbstractSearchRequest;
+use Perscom\Http\Requests\AbstractSearchRequest;
 
 class SearchSpecialtiesRequest extends AbstractSearchRequest
 {
