+++ src/Http/Requests/Submissions/SearchSubmissionsRequest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Perscom\Http\Requests\Submissions;
 
-use Perscom\RequestType\AbstractSearchRequest;
+use Perscom\Http\Requests\AbstractSearchRequest;
 
 class SearchSubmissionsRequest extends AbstractSearchRequest
 {
