+++ src/Http/Requests/Tasks/SearchTasksRequest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Perscom\Http\Requests\Tasks;
 
-use Perscom\RequestType\AbstractSearchRequest;
+use Perscom\Http\Requests\AbstractSearchRequest;
 
 class SearchTasksRequest extends AbstractSearchRequest
 {
